+++ Gruntfile.js
@@ -19,7 +19,7 @@ module.exports = function(grunt) {
 		distExports: 'dist',
 		moduleIn: 'JourneyCore',
 		moduleOut: 'glympse-journey-core',
-		moduleVersion: '1.0.1',
+		moduleVersion: '1.1.0',
 		buildsDir: 'builds'
       }
     }