+++ grunt/eslint.js
@@ -2,5 +2,7 @@ module.exports = {
   options: {
     config: '.eslintrc'
   },
-  target: ['<%= config.app %>/src/**/**.js']
+  target: ['<%= config.app %>/src/**/**.js'
+		  , '!<%= config.app %>/src/ph/**/**.js'
+		  ]
 };