+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 #Glympse Journey (Core) - Changelog
 ===================================
 
+
+1.1.0 [2015/07/26]: Remove demo dependencies, build cleanup, support 'future' eta type
+1.0.1/2 [2015/07/16]: Bower cfg updates (no build)
 1.0.0 [2015/07/15]: Initial release